+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Yii Framework 2 gii extension Change Log
 -----------------------
 
 - Enh #287: Model generator is now generating relation's phpdoc hints with target ActiveQuery class (bscheshirwork)
+- Bug #422: Fix relational query getter documentation style (mikk150)
 
 
 2.1.3 November 19, 2019