+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 Yii Framework 2 gii extension Change Log
 ========================================
 
-2.1.4 under development
------------------------
+2.1.4 January 17, 2020
+----------------------
 
-- Enh #287: Model generator is now generating relation's phpdoc hints with target ActiveQuery class (bscheshirwork)
 - Bug #422: Fix relational query getter documentation style (mikk150)
+- Enh #287: Model generator is now generating relation's phpdoc hints with target ActiveQuery class (bscheshirwork)
 
 
 2.1.3 November 19, 2019